// src/features/pretest/pretest-options.generated.ts
export type BilingualOption = { value: string; labelZh: string; labelEn: string };
// ...
export const ethnicityOptions: BilingualOption[] = [
  {
    value: "not_chinese_citizen",
    labelZh: "我不是中国公民",
    labelEn: "I am not a Chinese citizen",
  },
  {
    value: "han_zu",
    labelZh: "汉族",
    labelEn: "Han",
  },
  {
    value: "zhuang_zu",
    labelZh: "壮族",
    labelEn: "Zhuang",
  },
  {
    value: "man_zu",
    labelZh: "满族",
    labelEn: "Manchu",
  },
  {
    value: "hui_zu",
    labelZh: "回族",
    labelEn: "Hui",
  },
  {
    value: "miao_zu",
    labelZh: "苗族",
    labelEn: "Miao",
  },
  {
    value: "wei_wu_er_zu",
    labelZh: "维吾尔族",
    labelEn: "Uyghur",
  },
  {
    value: "tu_jia_zu",
    labelZh: "土家族",
    labelEn: "Tujia",
  },
  {
    value: "yi_zu",
    labelZh: "彝族",
    labelEn: "Yi",
  },
  {
    value: "meng_gu_zu",
    labelZh: "蒙古族",
    labelEn: "Mongol",
  },
  {
    value: "zang_zu",
    labelZh: "藏族",
    labelEn: "Tibetan",
  },
  {
    value: "bu_yi_zu",
    labelZh: "布依族",
    labelEn: "Bouyei",
  },
  {
    value: "dong_zu",
    labelZh: "侗族",
    labelEn: "Dong",
  },
  {
    value: "yao_zu",
    labelZh: "瑶族",
    labelEn: "Yao",
  },
  {
    value: "chao_xian_zu",
    labelZh: "朝鲜族",
    labelEn: "Korean",
  },
  {
    value: "bai_zu",
    labelZh: "白族",
    labelEn: "Bai",
  },
  {
    value: "ha_ni_zu",
    labelZh: "哈尼族",
    labelEn: "Hani",
  },
  {
    value: "ha_sa_ke_zu",
    labelZh: "哈萨克族",
    labelEn: "Kazakh",
  },
  {
    value: "li_zu",
    labelZh: "黎族",
    labelEn: "Li",
  },
  {
    value: "dai_zu",
    labelZh: "傣族",
    labelEn: "Dai",
  },
  {
    value: "she_zu",
    labelZh: "畲族",
    labelEn: "She",
  },
  {
    value: "li_su_zu",
    labelZh: "傈僳族",
    labelEn: "Lisu",
  },
  {
    value: "ge_lao_zu",
    labelZh: "仡佬族",
    labelEn: "Gelao",
  },
  {
    value: "dong_xiang_zu",
    labelZh: "东乡族",
    labelEn: "Dongxiang",
  },
  {
    value: "gao_shan_zu",
    labelZh: "高山族",
    labelEn: "Gaoshan",
  },
  {
    value: "la_hu_zu",
    labelZh: "拉祜族",
    labelEn: "Lahu",
  },
  {
    value: "shui_zu",
    labelZh: "水族",
    labelEn: "Sui",
  },
  {
    value: "wa_zu",
    labelZh: "佤族",
    labelEn: "Va",
  },
  {
    value: "na_xi_zu",
    labelZh: "纳西族",
    labelEn: "Naxi",
  },
  {
    value: "qiang_zu",
    labelZh: "羌族",
    labelEn: "Qiang",
  },
  {
    value: "tu_zu",
    labelZh: "土族",
    labelEn: "Tu",
  },
  {
    value: "mu_lao_zu",
    labelZh: "仫佬族",
    labelEn: "Mulao",
  },
  {
    value: "xi_bo_zu",
    labelZh: "锡伯族",
    labelEn: "Xibe",
  },
  {
    value: "ke_er_ke_zi_zu",
    labelZh: "柯尔克孜族",
    labelEn: "Kyrgyz",
  },
  {
    value: "da_wo_er_zu",
    labelZh: "达斡尔族",
    labelEn: "Daur",
  },
  {
    value: "jing_po_zu",
    labelZh: "景颇族",
    labelEn: "Jingpo",
  },
  {
    value: "mao_nan_zu",
    labelZh: "毛南族",
    labelEn: "Maonan",
  },
  {
    value: "sa_la_zu",
    labelZh: "撒拉族",
    labelEn: "Salar",
  },
  {
    value: "bu_lang_zu",
    labelZh: "布朗族",
    labelEn: "Blang",
  },
  {
    value: "ta_ji_ke_zu",
    labelZh: "塔吉克族",
    labelEn: "Tajik",
  },
  {
    value: "a_chang_zu",
    labelZh: "阿昌族",
    labelEn: "Achang",
  },
  {
    value: "pu_mi_zu",
    labelZh: "普米族",
    labelEn: "Pumi",
  },
  {
    value: "e_wen_ke_zu",
    labelZh: "鄂温克族",
    labelEn: "Ewenki",
  },
  {
    value: "nu_zu",
    labelZh: "怒族",
    labelEn: "Nu",
  },
  {
    value: "jing_zu",
    labelZh: "京族",
    labelEn: "Gin",
  },
  {
    value: "ji_nuo_zu",
    labelZh: "基诺族",
    labelEn: "Jino",
  },
  {
    value: "de_ang_zu",
    labelZh: "德昂族",
    labelEn: "De’ang",
  },
  {
    value: "bao_an_zu",
    labelZh: "保安族",
    labelEn: "Bonan",
  },
  {
    value: "e_luo_si_zu",
    labelZh: "俄罗斯族",
    labelEn: "Russian",
  },
  {
    value: "yu_gu_zu",
    labelZh: "裕固族",
    labelEn: "Yugur",
  },
  {
    value: "wu_zi_bie_ke_zu",
    labelZh: "乌兹别克族",
    labelEn: "Uzbek",
  },
  {
    value: "men_ba_zu",
    labelZh: "门巴族",
    labelEn: "Monba",
  },
  {
    value: "e_lun_chun_zu",
    labelZh: "鄂伦春族",
    labelEn: "Oroqen",
  },
  {
    value: "du_long_zu",
    labelZh: "独龙族",
    labelEn: "Derung",
  },
  {
    value: "ta_ta_er_zu",
    labelZh: "塔塔尔族",
    labelEn: "Tatar",
  },
  {
    value: "he_zhe_zu",
    labelZh: "赫哲族",
    labelEn: "Hezhen",
  },
  {
    value: "luo_ba_zu",
    labelZh: "珞巴族",
    labelEn: "Lhoba",
  },
];